// src/runtime/types/ui.ts
export type UiColor =
  | "primary"
  | "secondary"
  | "success"
  | "info"
  | "warning"
  | "error"
  | "neutral";

export type ButtonVariant =
  | "solid"
  | "outline"
  | "soft"
  | "ghost"
  | "link"
  | "subtle";

export type Rounded =
  | "rounded-none"
  | "rounded-sm"
  | "rounded"
  | "rounded-md"
  | "rounded-lg"
  | "rounded-xl"
  | "rounded-2xl"
  | "rounded-3xl"
  | "rounded-full";